When there are lots of information being displayed on the same screen it is diffcult for a trader to identify the information data, and hence this in the Graph.tsx file, which shows a linear line when there is a huge difference in the stock prices.
We introduced the ratio field since we no longer want to discriminate
between two equities and instead desire to track their ratios.
We've added similar fields since we also wanted to track
upper bound, lower bound, and the point at which these boundaries
are breached, called trigger alert.
In order to determine the ratio, price abc and price def were
finally included. We won't be setting up the graph to display them.
We require a timestamp field because we are recording everything in
relation to time.

The type of graph we wish to use to show the data is called a "view."
The term "column-pivots" used to exist and was what allowed us to
separate or distinguish between the stocks of ABC and DEF,
it is initially set to y line. Due to our concern with the ratios
between the two equities rather than their individual values,
we eliminated this. Our x-axis is taken care of by "row-pivots."
As a result, we can map each data point according to its timestamp.
The x-axis would be empty absent this. We may concentrate on a specific
area of a datapoint's data along the y-axis thanks to "columns."

Finally, we have to make a slight update to the componentDidUpdate
method.The change we want to make is on the argument we put in
this.table.update

Modifications in the DataManipulator.ts file
The Row interface is the first thing we need to change in this file.
The Row interface must be modified to conform to the new schema because
it was previously virtually identical to the old model in Graph.tsx
before we altered it.This modification is required because the
generateRow method returns an object, and that object's
structure must match the table's schema in the Graph component.
Finally, we have to update the generateRow function of the
DataManipulator class to properly process raw server data.